+++ src/query/sql/src/planner/plans/scalar.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::HashSet;
 use std::hash::Hash;
 
 use common_ast::ast::BinaryOperator;
@@ -37,7 +38,7 @@ pub trait ScalarExpr {
 
     fn is_deterministic(&self) -> bool;
 
-    fn used_tables(&self) -> TableSet;
+    fn used_tables(&self) -> HashSet<String>;
 
     // TODO: implement this in the future
     // fn outer_columns(&self) -> ColumnSet;
@@ -105,7 +106,7 @@ impl ScalarExpr for Scalar {
         }
     }
 
-    fn used_tables(&self) -> TableSet {
+    fn used_tables(&self) -> HashSet<String> {
         match self {
             Scalar::BoundColumnRef(scalar) => scalar.used_tables(),
             Scalar::ConstantExpr(scalar) => scalar.used_tables(),
@@ -303,11 +304,10 @@ impl ScalarExpr for BoundColumnRef {
         true
     }
 
-    fn used_tables(&self) -> TableSet {
-        let mut table_set = TableSet::new();
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
         if let Some(table_name) = &self.column.table_name {
-            dbg!(table_name);
-            table_set.insert(table_name.parse::<usize>().unwrap());
+            table_set.insert(table_name.to_string());
             return table_set;
         }
         table_set
@@ -334,8 +334,8 @@ impl ScalarExpr for ConstantExpr {
         true
     }
 
-    fn used_tables(&self) -> TableSet {
-        TableSet::new()
+    fn used_tables(&self) -> HashSet<String> {
+        HashSet::new()
     }
 }
 
@@ -361,8 +361,8 @@ impl ScalarExpr for AndExpr {
         self.left.is_deterministic() && self.right.is_deterministic()
     }
 
-    fn used_tables(&self) -> TableSet {
-        let mut table_set = TableSet::new();
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
         table_set.extend(self.left.used_tables());
         table_set.extend(self.right.used_tables());
         table_set
@@ -391,8 +391,11 @@ impl ScalarExpr for OrExpr {
         self.left.is_deterministic() && self.right.is_deterministic()
     }
 
-    fn used_tables(&self) -> TableSet {
-        todo!()
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
+        table_set.extend(self.left.used_tables());
+        table_set.extend(self.right.used_tables());
+        table_set
     }
 }
 
@@ -475,8 +478,8 @@ impl ScalarExpr for ComparisonExpr {
             && self.right.is_deterministic()
     }
 
-    fn used_tables(&self) -> TableSet {
-        let mut table_set = TableSet::new();
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
         table_set.extend(self.left.used_tables());
         table_set.extend(self.right.used_tables());
         table_set
@@ -511,8 +514,8 @@ impl ScalarExpr for AggregateFunction {
         false
     }
 
-    fn used_tables(&self) -> TableSet {
-        let mut table_set = TableSet::new();
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
         for scalar in self.args.iter() {
             table_set.extend(scalar.used_tables());
         }
@@ -549,8 +552,12 @@ impl ScalarExpr for FunctionCall {
             && self.arguments.iter().all(|arg| arg.is_deterministic())
     }
 
-    fn used_tables(&self) -> TableSet {
-        todo!()
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
+        for scalar in self.arguments.iter() {
+            table_set.extend(scalar.used_tables());
+        }
+        table_set
     }
 }
 
@@ -574,8 +581,8 @@ impl ScalarExpr for CastExpr {
         self.argument.is_deterministic()
     }
 
-    fn used_tables(&self) -> TableSet {
-        let mut table_set = TableSet::new();
+    fn used_tables(&self) -> HashSet<String> {
+        let mut table_set = HashSet::new();
         table_set.extend(self.argument.used_tables());
         table_set
     }
@@ -640,8 +647,8 @@ impl ScalarExpr for SubqueryExpr {
         false
     }
 
-    fn used_tables(&self) -> TableSet {
+    fn used_tables(&self) -> HashSet<String> {
         // Don't care used tables in subquery
-        TableSet::new()
+        HashSet::new()
     }
 }
